Moved function to v2.
